requirements.txt
# Core platform dependencies
python-dotenv>=1.0.0
pydantic>=2.8
pydantic-settings>=2.2
boto3>=1.30.0

# LangChain + ingestion
langchain>=0.2.0
langchain-core>=0.1.37
langchain-community>=0.2.1
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
PyMuPDF>=1.23.0
pandas>=2.1.0
duckdb>=1.4.4
pyarrow>=13.0
python-snappy>=0.6.1
unstructured[all-docs]>=0.7.3
openai>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0

# LLM clients and apps requested
ollama>=0.1.0
qwen-agent>=0.0.34
streamlit>=1.26.0
